pyproject.toml
[project]
name = "retrackt"
version = "0.1.0"
description = "Trace and Report Non Consensual Images on Reddit"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "streamlit>=1.49.1",
    "deepface>=0.0.95",
    "tensorflow>=2.10.1,<2.21.0",
    "tf-keras",
    "opencv-python-headless",
    "requests",
    "praw",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = []
